worldzero-mvp/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
authlib==1.3.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
httpx==0.27.0
itsdangerous==2.2.0
starlette==0.37.2
